Change font color, add max width
