+++ src/components/Homepage/OpenCollective/index.module.scss
@@ -3,5 +3,8 @@
         flex-flow: row wrap;
         justify-content: center;
         display: flex;
+        max-width: 920px;
+        margin: 0 auto;
+        padding: 0 20px;
     }
 }
